refactor: Class improved to abstract even more logic for the views.
- The abstract class now integrates functions and fields to be used on a mvvm pattern.
- New hook methods for observe the movel view and set view listeners.
- Added Timber.
